Fix wrong replacement of v-else during migration
Panel was migrated in 530074fb933a9fb2a17815f9521d96fbffc3b244 and
Dropdown was migrated in b188a8a108e4f91ac9f6201c9e14a86653c1a6a1.

We replaced v-else with v-if in line 18 for Panel, and line 10 for
Dropdown. This is because the v-else attribute had no corresponding v-if
pair, and thus causes a compilation error in Vue 2.0.

It did not cause a compilation error in Vue 1.0, so it was mistakenly
assumed that the syntax was accepted in Vue 1.0, and was working as
intended (even if it comes across as a rather "hackish" approach).

However, we discovered that the dangling v-else actually has no effect
even in Vue 1.0. It so happens that Vue 2.0's compiler got smarter and
was able to spot the error.

Let's restore the behavior of the original logic, by removing the
current v-if attribute that replaced the original v-else attribute.