Upgrade Bootstrap from v3 to v4
The following command is used to upgrade to v4:

    npm install bootstrap@4.1.1 --save

Note that this does not affect the final build of vue-strap, as the
build does not bundle Bootstrap, and the users of vue-strap must supply
their own Bootstrap files.

Therefore, the sole purpose of upgrading to v4 here is to facilitate
easier testing of the components inside this repository when migrating
to v4.