index.js: Change Vue instance to use a different element
I primarily suspect that using 'body' html tag caused problems because
the scripts (such as Vue) are also inside the <body> tag, which is
interfering with the actual script.

Moving the script tags out of the initialized Vue's instance (like for
example, using #mainContent as in the code), it seems to work fine.