Typeahead: Add right-alignment options for dropdown menu
The searchbar is positioned in the extreme right of the navbar in the
MarkBind website. This is a common position used by many websites.

When positioned in the extreme right, the dropdown menu gets cut off on
the right side if it has a big width, as the menu is aligned to the left
of the searchbar's textbox.

Let's fix this issue by providing the user with the choice of making the
dropdown menu right-aligned instead of the default alignment
(left-aligned).