Dropdown: Add pointer cursor to dropdown-toggle class
When isLi == true, the <a> is used instead of a <button> for the dropdown
button.

The pointer cursor is only available when the dropdown button is
<button>. When <a> is used (e.g. when used inside Navbar and Tabs), the
pointer cursor does not show, instead a normal cursor is shown.

Let's make the cursor consistent by adding a cursor pointer to
'dropdown-toggle', so that all dropdowns will have the cursor regardless
of whether it uses a <a> or a <button>.