TipBox: Use FontAwesome icons instead of emojis
The icons used inside TipBox are all emojis.

Emojis are browser-based and OS-based, causing different appearance of
the same emoji on different platforms. This make it hard to determine if
the emoji used is the right choice or not.

Let's swap the icons to FontAwesome's version. Their appearances are
consistent throughout all browsers and OSes.