Panel: Change header markdown rendering to non-inline
renderInline() is used when trying to render markdown syntax surrounded
by other normal text.

Some Markdown syntax are not supported inline. For example, one will not
expect a "header" to appear in the middle of other text, so Markdown
will not parse headers even if it is correctly written.

Change calls of md.renderInline() to md.render() for header rendering.