popoverMixins.js: Fix popovers going off screens in some scenarios
We are unable to find the exact reason why popovers go off the screen
when used in several scenarios as outlined in MarkBind/markbind#338.

This is a temporary fix. We just set the position of the popover
relative to the position of the mouse cursor. The trigger position is
now completely ignored.

This can be removed once we are able to figure out the issue behind the
original code that was using the trigger position to calculate the
popover position.