Panel: Move caret to its own wrapper
The caret is combined with the header content. It is usually rendered
like this:

> Some heading

However, custom header contents, after markdown rendering, will be
wrapped with block HTML tags such as <p> and <h1>, which will make
browsers render a line break between the caret and header content:

>
Some heading

Let's fix this by moving the caret to a separate div. Then, make
everything inside a panel header be on a single line by using CSS
'display: inline-block' and 'width: ...', such that the width adds up to
100%.

    +--------+-----------------+---------+
    | caret  |  header-content | buttons |
    | (32px) | (100% - 32 - 32)|  (32px) |
    +--------+-----------------+---------+

Thanks to @nusjzx for providing the fix to the button-wrapper in PR
https://github.com/MarkBind/vue-strap/pull/86, in which his use of
'calc(100% - 32px)' provided a source of inspiration for this fix. His
fix in that PR is incorporated into this commit because without it, the
responsive design will not work.

An alternative fix considered was to use md.renderInline(). However,
md.renderInline() does not render markdown headings.

Another alternative fix was to insert the caret directly inside the
header content. However, after PR #74, #81 and #83 were merged, new
problems continue to emerge.

These problems are reported as comments in these pages:
- https://github.com/MarkBind/vue-strap/pull/74
- https://github.com/MarkBind/vue-strap/pull/81
- https://github.com/MarkBind/vue-strap/pull/83
- https://github.com/MarkBind/markbind/issues/373
- https://github.com/MarkBind/markbind/issues/383

As such, the PRs for this alternative fix has been reverted by the
previous commits before this commit.