Panel: Fix popup button always showing when on seamless
For seamless panels, header buttons should only show when the user
hovers over the title.

However, the popup button is always shown to the user regardless of
whether the user hovers over it or not.

Let's fix the popup button showing behavior so that it will only show
when the user hovers over it.