Panel: Use @sindresorhus/slugify to generate anchor ID for headings
The main MarkBind project has switched to using the
@sindresorhus/slugify library[1] to generate the anchor IDs for
headings. The switch is done in one of the recently merged PR in the
main MarkBind project[2].

Our panels still uses the old anchor ID generation algorithm for
headings in its header.

Let's switch the anchor ID generation algorithm of panels to use
@sindresorhus/slugify.

[1] - https://github.com/sindresorhus/slugify
[2] - https://github.com/MarkBind/markbind/pull/578