Install uglifyjs-webpack-plugin v1.3.0
Running `npm run build` will cause UglifyJS to complain with the
following error:

    ERROR in ./dist/vue-strap.min.js from UglifyJs
    SyntaxError: Unexpected token: operator (>) [./dist/vue-strap.min.js:27219,28]

This is because the recently installed library @sindresorhus/slugify[1]
is an ES6 library, which uses arrow functions (=>), that is only
supported in ES6 onwards.

However, the UglifyJs that is incorporated into Webpack does not support
ES6[2]. Therefore, it does not recognize arrow functions.

Let's upgrade our version of UglifyJs to uglifyjs-webpack-plugin@v1.3.0,
which supports ES6.

At the time of writing, UglifyJs v2.1.1 is also available. However, that
requires a newer version of Webpack (v4.0.0), which we have not upgraded
to yet.

[1] - https://github.com/sindresorhus/slugify
[2] - https://github.com/mishoo/UglifyJS2/blob/v2.x/README.md#note