Run `npm install` with npm 5.8.0 (#100)
Whenever a new package is installed, the developer responsible for
installing the new package have to update package-lock.json.

The format for package-lock.json is slightly different for different
versions of npm. Developers do not always use the same version of npm,
so different parts of package-lock.json may be generated by different
npm versions.

Over time, the inconsistency in the format of package-lock.json makes
it difficult to maintain (especially when npm needs to be upgraded).

The team have decided to stick to npm 5.8.0.

Let's update package-lock.json to npm 5.8.0.