Remove position fixed classes from Navbar (#101)
Since MarkBind is adopting Holy Grail Layout, position:fixed
elements do not behave well. Let's remove it from navbar.

Since the dropdown menu of the typeahead search bar (which
is commonly used by navbars) overlaps with the page navigation,
it requires a sufficiently high z-index such that it is properly
placed in front of it, so let's also increase the typeahead search
bar's z-index.