Prevent double-fetching of panel with src and expanded
When such a panel is rendered, two ajax requests are fired.
One due to mounted() calling the retriever's fetch,
and another due to the watch listener being triggered by the setting
of localExpanded in created() which
makes another call to the retriever's fetch.

Since the retriever element is conditionally rendered if
(it has been open at least once) or (it is preloaded),
let's remove the delay attribute on the retriever element,
as now we can be sure that when the retriever element is needed,
we can fetch the data immediately.

This also lets us remove the redundant code that manually fetches
the retriever's content.