Delay popover close when trigger is hover
When a popover is triggered by by hovering, it is near impossible
for the user to mouse over the popover to select/copy or otherwise
interact with it.

Let's add a timeout to the closing of the popover on mouseleave only
when it is opened by a mouseenter. This timeout will be cancelled once
a mouseenter event is triggered again, so the popover stays open.

To allow the hovering over of a popover without disappearing,
authors can manually set the content of the popover to trigger
itself:

<popover id="hoverable">
  <div slot="content">
    <trigger for="hoverable">
      Hover over me to keep me open
    </trigger>
  </div>
  Hover to trigger
</popover>