Bump node-sass version to v4.0.0 (#121)
node-sass@3.10.1 does not build on nodejs 12.14.1 (LTS). The package is 
too old, so it uses features of nodejs that has been made obsolete,
resulting in a failed build when the package is building its C++
libraries.

Let's upgrade to node-sass@4.0.0. A simple run of `npm build` seems to
indicate that there's no changes to our `dist` builds, hence we should 
not expect any broken changes.